include tinymce html editor product

+++ database/migrations/2021_07_22_195325_create_products_table.php
@@ -13,24 +13,27 @@ class CreateProductsTable extends Migration
      */
     public function up()
     {
-        Schema::create('products', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('slug')->unique();
-            $table->text('short_description')->nullable();
-            $table->longText('description');
-            $table->decimal('regular_price');
-            $table->decimal('sale_price')->nullable();
-            $table->string('SKU');
-            $table->enum('stock_status', ['instock', 'outofstock']);
-            $table->boolean('featured')->default(false);
-            $table->unsignedInteger('quantity')->default(10);
-            $table->string('image')->nullable();
-            $table->text('images')->nullable();
-            $table->bigInteger('category_id')->unsigned()->nullable();
-            $table->timestamps();
-            $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
-        });
+        if ( ! Schema::hasTable('products')) {
+            Schema::create('products', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('slug')->unique();
+                $table->text('short_description')->nullable();
+                $table->text('additional_info')->nullable();
+                $table->longText('description');
+                $table->decimal('regular_price');
+                $table->decimal('sale_price')->nullable();
+                $table->string('SKU');
+                $table->enum('stock_status', ['instock', 'outofstock']);
+                $table->boolean('featured')->default(false);
+                $table->unsignedInteger('quantity')->default(10);
+                $table->string('image')->nullable();
+                $table->text('images')->nullable();
+                $table->bigInteger('category_id')->unsigned()->nullable();
+                $table->timestamps();
+                $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -40,6 +43,11 @@ class CreateProductsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('products');
+        // Schema::dropIfExists('products');
+        if ( ! Schema::hasColumn('products', 'additional_info')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->text('additional_info')->nullable()->after('description');
+                });
+        }
     }
 }
